+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/DefaultTrackSelector.java
@@ -43,7 +43,6 @@ import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.MediaSource.MediaPeriodId;
 import com.google.android.exoplayer2.source.TrackGroup;
 import com.google.android.exoplayer2.source.TrackGroupArray;
-import com.google.android.exoplayer2.trackselection.TrackSelectionOverrides.TrackSelectionOverride;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.BundleableUtil;
 import com.google.android.exoplayer2.util.MimeTypes;
@@ -593,9 +592,32 @@ public class DefaultTrackSelector extends MappingTrackSelector {
     }
 
     @Override
-    public ParametersBuilder setTrackSelectionOverrides(
-        TrackSelectionOverrides trackSelectionOverrides) {
-      super.setTrackSelectionOverrides(trackSelectionOverrides);
+    public ParametersBuilder addOverride(TrackSelectionOverride override) {
+      super.addOverride(override);
+      return this;
+    }
+
+    @Override
+    public ParametersBuilder clearOverride(TrackGroup trackGroup) {
+      super.clearOverride(trackGroup);
+      return this;
+    }
+
+    @Override
+    public ParametersBuilder setOverrideForType(TrackSelectionOverride override) {
+      super.setOverrideForType(override);
+      return this;
+    }
+
+    @Override
+    public ParametersBuilder clearOverridesOfType(@C.TrackType int trackType) {
+      super.clearOverridesOfType(trackType);
+      return this;
+    }
+
+    @Override
+    public ParametersBuilder clearOverrides() {
+      super.clearOverrides();
       return this;
     }
 
@@ -700,7 +722,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      * @param groups The {@link TrackGroupArray} for which the override should be applied.
      * @param override The override.
      * @return This builder.
-     * @deprecated Use {@link TrackSelectionParameters.Builder#setTrackSelectionOverrides}.
+     * @deprecated Use {@link TrackSelectionParameters.Builder#addOverride(TrackSelectionOverride)}.
      */
     @Deprecated
     public final ParametersBuilder setSelectionOverride(
@@ -725,7 +747,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      * @param rendererIndex The renderer index.
      * @param groups The {@link TrackGroupArray} for which the override should be cleared.
      * @return This builder.
-     * @deprecated Use {@link TrackSelectionParameters.Builder#setTrackSelectionOverrides}.
+     * @deprecated Use {@link TrackSelectionParameters.Builder#clearOverride(TrackGroup)}.
      */
     @Deprecated
     public final ParametersBuilder clearSelectionOverride(
@@ -748,7 +770,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      *
      * @param rendererIndex The renderer index.
      * @return This builder.
-     * @deprecated Use {@link TrackSelectionParameters.Builder#setTrackSelectionOverrides}.
+     * @deprecated Use {@link TrackSelectionParameters.Builder#clearOverridesOfType(int)}.
      */
     @Deprecated
     public final ParametersBuilder clearSelectionOverrides(int rendererIndex) {
@@ -766,7 +788,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      * Clears all track selection overrides for all renderers.
      *
      * @return This builder.
-     * @deprecated Use {@link TrackSelectionParameters.Builder#setTrackSelectionOverrides}.
+     * @deprecated Use {@link TrackSelectionParameters.Builder#clearOverrides()}.
      */
     @Deprecated
     public final ParametersBuilder clearSelectionOverrides() {
@@ -1017,7 +1039,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      * @return Whether there is an override.
      * @deprecated Only works to retrieve the overrides set with the deprecated {@link
      *     ParametersBuilder#setSelectionOverride(int, TrackGroupArray, SelectionOverride)}. Use
-     *     {@link TrackSelectionParameters#trackSelectionOverrides} instead.
+     *     {@link TrackSelectionParameters#overrides} instead.
      */
     @Deprecated
     public final boolean hasSelectionOverride(int rendererIndex, TrackGroupArray groups) {
@@ -1034,7 +1056,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
      * @return The override, or null if no override exists.
      * @deprecated Only works to retrieve the overrides set with the deprecated {@link
      *     ParametersBuilder#setSelectionOverride(int, TrackGroupArray, SelectionOverride)}. Use
-     *     {@link TrackSelectionParameters#trackSelectionOverrides} instead.
+     *     {@link TrackSelectionParameters#overrides} instead.
      */
     @Deprecated
     @Nullable
@@ -1653,9 +1675,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
       TrackGroupArray rendererTrackGroups = mappedTrackInfo.getTrackGroups(rendererIndex);
       for (int j = 0; j < rendererTrackGroups.length; j++) {
         maybeUpdateApplicableOverrides(
-            applicableOverrides,
-            params.trackSelectionOverrides.getOverride(rendererTrackGroups.get(j)),
-            rendererIndex);
+            applicableOverrides, params.overrides.get(rendererTrackGroups.get(j)), rendererIndex);
       }
     }
     // Also iterate unmapped groups to see if they have overrides.
@@ -1663,7 +1683,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
     for (int i = 0; i < unmappedGroups.length; i++) {
       maybeUpdateApplicableOverrides(
           applicableOverrides,
-          params.trackSelectionOverrides.getOverride(unmappedGroups.get(i)),
+          params.overrides.get(unmappedGroups.get(i)),
           /* rendererIndex= */ C.INDEX_UNSET);
     }
     return applicableOverrides;