+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/StyledPlayerControlView.java
@@ -67,8 +67,7 @@ import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.TracksInfo;
 import com.google.android.exoplayer2.TracksInfo.TrackGroupInfo;
 import com.google.android.exoplayer2.source.TrackGroup;
-import com.google.android.exoplayer2.trackselection.TrackSelectionOverrides;
-import com.google.android.exoplayer2.trackselection.TrackSelectionOverrides.TrackSelectionOverride;
+import com.google.android.exoplayer2.trackselection.TrackSelectionOverride;
 import com.google.android.exoplayer2.trackselection.TrackSelectionParameters;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.RepeatModeUtil;
@@ -1908,7 +1907,7 @@ public class StyledPlayerControlView extends FrameLayout {
       holder.textView.setText(R.string.exo_track_selection_auto);
       // hasSelectionOverride is true means there is an explicit track selection, not "Auto".
       TrackSelectionParameters parameters = checkNotNull(player).getTrackSelectionParameters();
-      boolean hasSelectionOverride = hasSelectionOverride(parameters.trackSelectionOverrides);
+      boolean hasSelectionOverride = hasSelectionOverride(parameters);
       holder.checkView.setVisibility(hasSelectionOverride ? INVISIBLE : VISIBLE);
       holder.itemView.setOnClickListener(
           v -> {
@@ -1917,12 +1916,6 @@ public class StyledPlayerControlView extends FrameLayout {
             }
             TrackSelectionParameters trackSelectionParameters =
                 player.getTrackSelectionParameters();
-            TrackSelectionOverrides trackSelectionOverrides =
-                trackSelectionParameters
-                    .trackSelectionOverrides
-                    .buildUpon()
-                    .clearOverridesOfType(C.TRACK_TYPE_AUDIO)
-                    .build();
             Set<@C.TrackType Integer> disabledTrackTypes =
                 new HashSet<>(trackSelectionParameters.disabledTrackTypes);
             disabledTrackTypes.remove(C.TRACK_TYPE_AUDIO);
@@ -1930,7 +1923,7 @@ public class StyledPlayerControlView extends FrameLayout {
                 .setTrackSelectionParameters(
                     trackSelectionParameters
                         .buildUpon()
-                        .setTrackSelectionOverrides(trackSelectionOverrides)
+                        .clearOverridesOfType(C.TRACK_TYPE_AUDIO)
                         .setDisabledTrackTypes(disabledTrackTypes)
                         .build());
             settingsAdapter.setSubTextAtPosition(
@@ -1940,10 +1933,10 @@ public class StyledPlayerControlView extends FrameLayout {
           });
     }
 
-    private boolean hasSelectionOverride(TrackSelectionOverrides trackSelectionOverrides) {
+    private boolean hasSelectionOverride(TrackSelectionParameters trackSelectionParameters) {
       for (int i = 0; i < tracks.size(); i++) {
         TrackGroup trackGroup = tracks.get(i).trackGroupInfo.getTrackGroup();
-        if (trackSelectionOverrides.getOverride(trackGroup) != null) {
+        if (trackSelectionParameters.overrides.containsKey(trackGroup)) {
           return true;
         }
       }
@@ -1966,7 +1959,7 @@ public class StyledPlayerControlView extends FrameLayout {
             getResources().getString(R.string.exo_track_selection_none));
         // TODO(insun) : Make the audio item in main settings (settingsAdapater)
         //  to be non-clickable.
-      } else if (!hasSelectionOverride(params.trackSelectionOverrides)) {
+      } else if (!hasSelectionOverride(params)) {
         settingsAdapter.setSubTextAtPosition(
             SETTINGS_AUDIO_TRACK_SELECTION_POSITION,
             getResources().getString(R.string.exo_track_selection_auto));
@@ -2017,8 +2010,7 @@ public class StyledPlayerControlView extends FrameLayout {
         TrackInformation track = tracks.get(position - 1);
         TrackGroup trackGroup = track.trackGroupInfo.getTrackGroup();
         TrackSelectionParameters params = checkNotNull(player).getTrackSelectionParameters();
-        boolean explicitlySelected =
-            params.trackSelectionOverrides.getOverride(trackGroup) != null && track.isSelected();
+        boolean explicitlySelected = params.overrides.get(trackGroup) != null && track.isSelected();
         holder.textView.setText(track.trackName);
         holder.checkView.setVisibility(explicitlySelected ? VISIBLE : INVISIBLE);
         holder.itemView.setOnClickListener(
@@ -2028,14 +2020,6 @@ public class StyledPlayerControlView extends FrameLayout {
               }
               TrackSelectionParameters trackSelectionParameters =
                   player.getTrackSelectionParameters();
-              TrackSelectionOverrides overrides =
-                  trackSelectionParameters
-                      .trackSelectionOverrides
-                      .buildUpon()
-                      .setOverrideForType(
-                          new TrackSelectionOverride(
-                              trackGroup, ImmutableList.of(track.trackIndex)))
-                      .build();
               Set<@C.TrackType Integer> disabledTrackTypes =
                   new HashSet<>(trackSelectionParameters.disabledTrackTypes);
               disabledTrackTypes.remove(track.trackGroupInfo.getTrackType());
@@ -2043,7 +2027,9 @@ public class StyledPlayerControlView extends FrameLayout {
                   .setTrackSelectionParameters(
                       trackSelectionParameters
                           .buildUpon()
-                          .setTrackSelectionOverrides(overrides)
+                          .setOverrideForType(
+                              new TrackSelectionOverride(
+                                  trackGroup, ImmutableList.of(track.trackIndex)))
                           .setDisabledTrackTypes(disabledTrackTypes)
                           .build());
               onTrackSelection(track.trackName);