+++ RELEASENOTES.md
@@ -56,6 +56,9 @@
         values in `Player.EventFlags`.
     *   Split `AnalyticsCollector` into an interface and default implementation
         to allow it to be stripped by R8 if an app doesn't need it.
+*   Track selection
+    *   Flatten `TrackSelectionOverrides` class into `TrackSelectionParameters`,
+        and promote `TrackSelectionOverride` to a top level class.
 *   Android 12 compatibility:
     *   Upgrade the Cast extension to depend on
         `com.google.android.gms:play-services-cast-framework:20.1.0`. Earlier